[Containerapp] Fix #29128: `az containerapp env certificate`: Support managed certificate (#29156)

+++ src/azure-cli/azure/cli/command_modules/containerapp/custom.py
@@ -3751,15 +3751,25 @@ def containerapp_up_logic(cmd, resource_group_name, name, managed_env, image, en
     return create_containerapp(cmd=cmd, name=name, resource_group_name=resource_group_name, managed_env=managed_env, image=image, env_vars=env_vars, ingress=ingress, target_port=target_port, registry_server=registry_server, registry_user=registry_user, registry_pass=registry_pass, workload_profile_name=workload_profile_name)
 
 
-def list_certificates(cmd, name, resource_group_name, location=None, certificate=None, thumbprint=None):
-    if certificate and is_valid_resource_id(certificate):
-        certificate_type = parse_resource_id(certificate)["resource_type"]
-    else:
-        certificate_type = PRIVATE_CERTIFICATE_RT
+def create_managed_certificate(cmd, name, resource_group_name, hostname, validation_method, certificate_name=None):
+    if certificate_name and not check_managed_cert_name_availability(cmd, resource_group_name, name, certificate_name):
+        raise ValidationError(f"Certificate name '{certificate_name}' is not available.")
+    cert_name = certificate_name
+    while not cert_name:
+        cert_name = generate_randomized_managed_cert_name(hostname, resource_group_name)
+        if not check_managed_cert_name_availability(cmd, resource_group_name, name, certificate_name):
+            cert_name = None
+    certificate_envelop = prepare_managed_certificate_envelop(cmd, name, resource_group_name, hostname, validation_method.upper())
+    try:
+        r = ManagedEnvironmentClient.create_or_update_managed_certificate(cmd, resource_group_name, name, cert_name, certificate_envelop, True, validation_method.upper() == 'TXT')
+        return r
+    except Exception as e:
+        handle_raw_exception(e)
 
-    if certificate_type != PRIVATE_CERTIFICATE_RT:
-        raise ValidationError(f"The certificate {certificate} is not private-key certificate.")
-    return list_certificates_logic(cmd, name, resource_group_name, location, certificate, thumbprint, private_key_certificates_only=True)
+
+def list_certificates(cmd, name, resource_group_name, location=None, certificate=None, thumbprint=None, managed_certificates_only=False, private_key_certificates_only=False):
+
+    return list_certificates_logic(cmd, name, resource_group_name, location, certificate, thumbprint, managed_certificates_only=managed_certificates_only, private_key_certificates_only=private_key_certificates_only)
 
 
 def list_certificates_logic(cmd, name, resource_group_name, location=None, certificate=None, thumbprint=None, managed_certificates_only=False, private_key_certificates_only=False):
@@ -3867,23 +3877,8 @@ def upload_certificate(cmd, name, resource_group_name, certificate_file, certifi
 
 
 def delete_certificate(cmd, resource_group_name, name, location=None, certificate=None, thumbprint=None):
-    if not certificate and not thumbprint:
-        raise RequiredArgumentMissingError('Please specify at least one of parameters: --certificate and --thumbprint')
-    # validate for GA
-    cert_name = certificate
-    if certificate and is_valid_resource_id(certificate):
-        cert_type = parse_resource_id(certificate)["resource_type"]
-        cert_name = parse_resource_id(certificate)["resource_name"]
-    else:
-        cert_type = PRIVATE_CERTIFICATE_RT
 
-    if thumbprint:
-        cert_type = PRIVATE_CERTIFICATE_RT
-
-    if cert_type != PRIVATE_CERTIFICATE_RT:
-        raise ValidationError(f"The certificate {cert_name} is not private-key certificate.")
-
-    delete_certificate_logic(cmd, resource_group_name, name, cert_name, location, certificate, thumbprint, cert_type)
+    delete_certificate_logic(cmd=cmd, resource_group_name=resource_group_name, name=name, cert_name=certificate, location=location, certificate=certificate, thumbprint=thumbprint)
 
 
 # this function will be used in extension
@@ -3923,11 +3918,19 @@ def delete_certificate_logic(cmd, resource_group_name, name, cert_name, location
             raise ResourceNotFoundError(f"The certificate '{cert_name}' does not exist in Container app environment '{name}'.")
         if len(managed_certs) > 0 and len(private_certs) > 0:
             raise RequiredArgumentMissingError(f"Found more than one certificates with name '{cert_name}':\n'{managed_certs[0]['id']}',\n'{private_certs[0]['id']}'.\nPlease specify the certificate id using --certificate.")
-        try:
-            ManagedEnvironmentClient.delete_managed_certificate(cmd, resource_group_name, name, cert_name)
-            logger.warning('Successfully deleted certificate: %s', cert_name)
-        except Exception as e:
-            handle_raw_exception(e)
+        if private_certs:
+            try:
+                ManagedEnvironmentClient.delete_certificate(cmd, resource_group_name, name, cert_name)
+                logger.warning('Successfully deleted certificate: %s', cert_name)
+            except Exception as e:
+                handle_raw_exception(e)
+
+        if managed_certs:
+            try:
+                ManagedEnvironmentClient.delete_managed_certificate(cmd, resource_group_name, name, cert_name)
+                logger.warning('Successfully deleted certificate: %s', cert_name)
+            except Exception as e:
+                handle_raw_exception(e)
 
 
 def upload_ssl(cmd, resource_group_name, name, environment, certificate_file, hostname, certificate_password=None, certificate_name=None, location=None):
@@ -3956,11 +3959,9 @@ def upload_ssl(cmd, resource_group_name, name, environment, certificate_file, ho
     return patch_new_custom_domain(cmd, resource_group_name, name, new_custom_domains)
 
 
-def bind_hostname(cmd, resource_group_name, name, hostname, thumbprint=None, certificate=None, location=None, environment=None):
-    # validate for GA
-    if not thumbprint and not certificate:
-        raise RequiredArgumentMissingError('Please specify at least one of parameters: --certificate and --thumbprint')
-    return bind_hostname_logic(cmd, resource_group_name, name, hostname, thumbprint, certificate, location, environment)
+def bind_hostname(cmd, resource_group_name, name, hostname, thumbprint=None, certificate=None, location=None, environment=None, validation_method=None):
+
+    return bind_hostname_logic(cmd, resource_group_name, name, hostname, thumbprint, certificate, location, environment, validation_method)
 
 
 def bind_hostname_logic(cmd, resource_group_name, name, hostname, thumbprint=None, certificate=None, location=None, environment=None, validation_method=None):
